+++ DeviceInfo/DeviceInfoImplementation.cpp
@@ -89,12 +89,12 @@ namespace Plugin {
 
     uint32_t DeviceInfoImplementation::ModelName(string& value) const  
     {
-        value= _modelName;
+        value = _modelName;
         return Core::ERROR_NONE;
     }
-    uint32_t DeviceInfoImplementation::ModelYear(uint32_t& value) const 
+    uint32_t DeviceInfoImplementation::ModelYear(uint16_t& value) const 
     {
-        value= _modelYear;
+        value = _modelYear;
         return Core::ERROR_NONE;
     }
     uint32_t DeviceInfoImplementation::FriendlyName(string& value) const 
@@ -104,7 +104,7 @@ namespace Plugin {
     }
     uint32_t DeviceInfoImplementation::SystemIntegratorName(string& value) const 
     {
-        value= _systemIntegratorName;
+        value = _systemIntegratorName;
         return Core::ERROR_NONE;
     }
     uint32_t DeviceInfoImplementation::PlatformName(string& value) const
